https://hackerone.com/reports/10529

+++ web/concrete/tools/users/bulk_delete.php
@@ -11,7 +11,9 @@ $tp = new TaskPermission();
 $u = new User();
 $sk = PermissionKey::getByHandle('access_user_search');
 $tp = new TaskPermission();
-if (!$tp->canDeleteUser()) { 
+$token = Loader::helper('validation/token');
+
+if (!$tp->canDeleteUser()) {
 	die(t("Access Denied."));
 }
 
@@ -23,8 +25,8 @@ $excluded_user_ids[] = USER_SUPER_ID; // can't delete the super user (admin)
 
 if (is_array($_REQUEST['uID'])) {
 	foreach($_REQUEST['uID'] as $uID) {
-		$ui = UserInfo::getByID($uID);		
-		if(!$sk->validate($ui) || (in_array($ui->getUserID(),$excluded_user_ids))) { 
+		$ui = UserInfo::getByID($uID);
+		if(!$sk->validate($ui) || (in_array($ui->getUserID(),$excluded_user_ids))) {
 			$excluded = true;
 		} else {
 			$users[] = $ui;
@@ -33,6 +35,11 @@ if (is_array($_REQUEST['uID'])) {
 }
 
 if ($_POST['task'] == 'delete') {
+	if (!$token->validate('bulk_user_delete')) {
+		echo Loader::helper('json')->encode(array('error'=>t("Invalid token.")));
+		exit;
+	}
+
 	foreach($users as $ui) {
 		if(!(in_array($ui->getUserID(),$excluded_user_ids))) {
 			$ui->delete();
@@ -40,7 +47,7 @@ if ($_POST['task'] == 'delete') {
 	}
 	echo Loader::helper('json')->encode(array('error'=>false));
 	exit;
-} 
+}
 
 if (!isset($_REQUEST['reload'])) { ?>
 	<div id="ccm-user-bulk-delete-wrapper">
@@ -49,6 +56,7 @@ if (!isset($_REQUEST['reload'])) { ?>
 	<div id="ccm-user-delete" class="ccm-ui">
 		<form method="post" id="ccm-user-bulk-delete" action="<?php echo REL_DIR_FILES_TOOLS_REQUIRED ?>/users/bulk_delete">
 			<?php
+			$token->output('bulk_user_delete');
 			echo $form->hidden('task','delete');
 			foreach($users as $ui) {
 				echo $form->hidden('uID[]' , $ui->getUserID());
@@ -63,7 +71,7 @@ if (!isset($_REQUEST['reload'])) { ?>
 		</form>
 	</div>
 	<div class="dialog-buttons">
-		<?=$ih->button_js(t('Cancel'), 'jQuery.fn.dialog.closeTop()', 'left', 'btn')?>	
+		<?=$ih->button_js(t('Cancel'), 'jQuery.fn.dialog.closeTop()', 'left', 'btn')?>
 		<?=$ih->button_js(t('Delete'), 'ccm_userBulkActivate()', 'right', 'btn error')?>
 	</div>
 <?
@@ -72,7 +80,7 @@ if (!isset($_REQUEST['reload'])) { ?>
 <? } ?>
 
 <script type="text/javascript">
-ccm_userBulkActivate = function() { 
+ccm_userBulkActivate = function() {
 	jQuery.fn.dialog.showLoader();
 	$("#ccm-user-bulk-delete").ajaxSubmit(function(resp) {
 		jQuery.fn.dialog.closeTop();
@@ -83,6 +91,6 @@ ccm_userBulkActivate = function() {
 		       ccm_parseAdvancedSearchResponse(r, '<?=$searchInstance?>');
 		});
 	});
-	
+
 };
 </script>
